chore: remove bundled model definitions
