+++ packages/plugin-commerce-api/src/commerce/collections/index.ts
@@ -1,11 +1,14 @@
 /* DTO */
+export * from "./dto/collection-seo.dto";
 export * from "./dto/create-collection.dto";
 export * from "./dto/update-collection.dto";
 export * from "./dto/collection-response.dto";
 
 /* Models */
 export * from "./models/collection-status.enum";
-export * from "./models/collection.models";
+export * from "./models/collection-seo.model";
+export * from "./models/collection-base.model";
+export * from "./models/collection-response.model";
 
 /* Schemas */
 export * from "./schemas/product-collection.schema";
