+++ packages/plugin-commerce-api/src/commerce/collections/models/collection-base.model.ts
@@ -1,0 +1,17 @@
+import type { CollectionStatus } from "./collection-status.enum";
+import type { CollectionSeoModel } from "./collection-seo.model";
+
+export type CollectionBaseModel = {
+  slug: string;
+  language: string;
+  status: CollectionStatus;
+  title: string;
+  description?: string;
+  seo?: CollectionSeoModel;
+  tags?: string[];
+  publishAt?: string;
+  expireAt?: string;
+  coverImage?: string;
+  parentId?: string;
+  sortOrder?: number;
+};
