+++ packages/plugin-commerce-api/src/commerce/collections/models/collection-response.model.ts
@@ -1,0 +1,7 @@
+export type CollectionResponseModel = {
+  id: string;
+  slugs: Record<string, string>;
+  translations: Record<string, Record<string, unknown>>;
+  createdAt: Date;
+  updatedAt: Date;
+};
