+++ packages/plugin-commerce-api/src/commerce/collections/models/collection-seo.model.ts
@@ -1,0 +1,6 @@
+export type CollectionSeoModel = {
+  metaTitle?: string;
+  metaDescription?: string;
+  metaKeywords?: string[];
+  canonicalUrl?: string;
+};
