+++ packages/plugin-commerce-api/src/commerce/customer-addresses/index.ts
@@ -4,5 +4,6 @@ export * from "./customer-addresses.controller";
 export * from "./dto/create-customer-address.dto";
 export * from "./dto/update-customer-address.dto";
 export * from "./dto/customer-address-response.dto";
-export * from "./models/customer-address.model";
+export * from "./models/customer-address-base.model";
+export * from "./models/customer-address-response.model";
 export * from "./schemas/customer-address.schema";
