+++ packages/plugin-commerce-api/src/commerce/customer-addresses/models/customer-address-base.model.ts
@@ -14,9 +14,3 @@ export type CustomerAddressBaseModel = {
   isDefaultShipping?: boolean;
   isDefaultBilling?: boolean;
 };
-
-export type CustomerAddressResponseModel = CustomerAddressBaseModel & {
-  id: string;
-  createdAt: Date;
-  updatedAt: Date;
-};
