+++ packages/plugin-commerce-api/src/commerce/customer-addresses/models/customer-address-response.model.ts
@@ -1,0 +1,7 @@
+import type { CustomerAddressBaseModel } from "./customer-address-base.model";
+
+export type CustomerAddressResponseModel = CustomerAddressBaseModel & {
+  id: string;
+  createdAt: Date;
+  updatedAt: Date;
+};
