+++ packages/plugin-commerce-api/src/commerce/orders/models/order-address.model.ts
@@ -1,0 +1,12 @@
+export type OrderAddressModel = {
+  firstName?: string;
+  lastName?: string;
+  company?: string;
+  address1: string;
+  address2?: string;
+  city: string;
+  postalCode?: string;
+  province?: string;
+  countryCode: string;
+  phone?: string;
+};
