+++ packages/plugin-commerce-api/src/commerce/orders/models/order-item.model.ts
@@ -1,0 +1,10 @@
+export type OrderItemModel = {
+  title: string;
+  variantTitle?: string;
+  quantity: number;
+  unitPrice: number;
+  currencyCode: string;
+  productId?: string;
+  variantId?: string;
+  sku?: string;
+};
