+++ packages/plugin-commerce-api/src/commerce/products/dto/create-product.dto.ts
@@ -2,102 +2,18 @@ import { ApiProperty, ApiPropertyOptional } from "@nestjs/swagger";
 import { Type } from "class-transformer";
 import {
   IsArray,
-  IsBoolean,
   IsDateString,
   IsEnum,
   IsNotEmpty,
-  IsNumber,
   IsOptional,
   IsString,
   ValidateNested,
 } from "class-validator";
+
 import { ProductStatus } from "../models/product-status.enum";
-import type {
-  ProductBaseModel,
-  ProductPriceModel,
-  ProductSeoModel,
-  ProductVariantModel,
-} from "../models/product.models";
-
-export class ProductSeoDto implements ProductSeoModel {
-  @ApiPropertyOptional()
-  @IsOptional()
-  @IsString()
-  metaTitle?: string;
-
-  @ApiPropertyOptional()
-  @IsOptional()
-  @IsString()
-  metaDescription?: string;
-
-  @ApiPropertyOptional({ type: [String] })
-  @IsOptional()
-  @IsArray()
-  @IsString({ each: true })
-  metaKeywords?: string[];
-
-  @ApiPropertyOptional()
-  @IsOptional()
-  @IsString()
-  canonicalUrl?: string;
-}
-
-export class ProductPriceDto implements ProductPriceModel {
-  @ApiProperty()
-  @IsNotEmpty()
-  @IsString()
-  currencyCode: string;
-
-  @ApiProperty()
-  @Type(() => Number)
-  @IsNumber()
-  amount: number;
-
-  @ApiPropertyOptional()
-  @IsOptional()
-  @Type(() => Number)
-  @IsNumber()
-  compareAtAmount?: number;
-}
-
-export class ProductVariantDto implements ProductVariantModel {
-  @ApiPropertyOptional()
-  @IsOptional()
-  @IsString()
-  id?: string;
-
-  @ApiProperty()
-  @IsNotEmpty()
-  @IsString()
-  title: string;
-
-  @ApiProperty()
-  @IsNotEmpty()
-  @IsString()
-  sku: string;
-
-  @ApiPropertyOptional()
-  @IsOptional()
-  @IsString()
-  barcode?: string;
-
-  @ApiPropertyOptional({ type: () => [ProductPriceDto] })
-  @IsOptional()
-  @ValidateNested({ each: true })
-  @Type(() => ProductPriceDto)
-  prices?: ProductPriceDto[];
-
-  @ApiPropertyOptional()
-  @IsOptional()
-  @Type(() => Number)
-  @IsNumber()
-  inventoryQuantity?: number;
-
-  @ApiPropertyOptional()
-  @IsOptional()
-  @IsBoolean()
-  allowBackorder?: boolean;
-}
+import type { ProductBaseModel } from "../models/product-base.model";
+import { ProductSeoDto } from "./product-seo.dto";
+import { ProductVariantDto } from "./product-variant.dto";
 
 export class CreateProductDto implements ProductBaseModel {
   @ApiProperty()
