+++ packages/plugin-commerce-api/src/commerce/products/dto/product-seo.dto.ts
@@ -1,0 +1,27 @@
+import { ApiPropertyOptional } from "@nestjs/swagger";
+import { IsArray, IsOptional, IsString } from "class-validator";
+
+import type { ProductSeoModel } from "../models/product-seo.model";
+
+export class ProductSeoDto implements ProductSeoModel {
+  @ApiPropertyOptional()
+  @IsOptional()
+  @IsString()
+  metaTitle?: string;
+
+  @ApiPropertyOptional()
+  @IsOptional()
+  @IsString()
+  metaDescription?: string;
+
+  @ApiPropertyOptional({ type: [String] })
+  @IsOptional()
+  @IsArray()
+  @IsString({ each: true })
+  metaKeywords?: string[];
+
+  @ApiPropertyOptional()
+  @IsOptional()
+  @IsString()
+  canonicalUrl?: string;
+}
