+++ packages/plugin-commerce-api/src/commerce/products/dto/product-variant.dto.ts
@@ -1,0 +1,52 @@
+import { ApiProperty, ApiPropertyOptional } from "@nestjs/swagger";
+import { Type } from "class-transformer";
+import {
+  IsBoolean,
+  IsNotEmpty,
+  IsNumber,
+  IsOptional,
+  IsString,
+  ValidateNested,
+} from "class-validator";
+
+import type { ProductVariantModel } from "../models/product-variant.model";
+import { ProductPriceDto } from "./product-price.dto";
+
+export class ProductVariantDto implements ProductVariantModel {
+  @ApiPropertyOptional()
+  @IsOptional()
+  @IsString()
+  id?: string;
+
+  @ApiProperty()
+  @IsNotEmpty()
+  @IsString()
+  title: string;
+
+  @ApiProperty()
+  @IsNotEmpty()
+  @IsString()
+  sku: string;
+
+  @ApiPropertyOptional()
+  @IsOptional()
+  @IsString()
+  barcode?: string;
+
+  @ApiPropertyOptional({ type: () => [ProductPriceDto] })
+  @IsOptional()
+  @ValidateNested({ each: true })
+  @Type(() => ProductPriceDto)
+  prices?: ProductPriceDto[];
+
+  @ApiPropertyOptional()
+  @IsOptional()
+  @Type(() => Number)
+  @IsNumber()
+  inventoryQuantity?: number;
+
+  @ApiPropertyOptional()
+  @IsOptional()
+  @IsBoolean()
+  allowBackorder?: boolean;
+}
