+++ packages/plugin-commerce-api/src/commerce/products/index.ts
@@ -1,8 +1,15 @@
 /* Models */
 export * from "./models/product-status.enum";
-export * from "./models/product.models";
+export * from "./models/product-seo.model";
+export * from "./models/product-price.model";
+export * from "./models/product-variant.model";
+export * from "./models/product-base.model";
+export * from "./models/product-response.model";
 
 /* DTO */
+export * from "./dto/product-seo.dto";
+export * from "./dto/product-price.dto";
+export * from "./dto/product-variant.dto";
 export * from "./dto/create-product.dto";
 export * from "./dto/update-product.dto";
 export * from "./dto/product-response.dto";
