+++ packages/plugin-commerce-api/src/commerce/products/models/product-price.model.ts
@@ -1,0 +1,5 @@
+export type ProductPriceModel = {
+  currencyCode: string;
+  amount: number;
+  compareAtAmount?: number;
+};
