+++ packages/plugin-commerce-api/src/commerce/products/models/product-response.model.ts
@@ -1,0 +1,7 @@
+export type ProductResponseModel = {
+  id: string;
+  slugs: Record<string, string>;
+  translations: Record<string, Record<string, unknown>>;
+  createdAt: Date;
+  updatedAt: Date;
+};
