+++ packages/plugin-commerce-api/src/commerce/products/models/product-seo.model.ts
@@ -1,0 +1,6 @@
+export type ProductSeoModel = {
+  metaTitle?: string;
+  metaDescription?: string;
+  metaKeywords?: string[];
+  canonicalUrl?: string;
+};
