+++ packages/plugin-commerce-api/src/commerce/products/models/product-variant.model.ts
@@ -1,0 +1,11 @@
+import type { ProductPriceModel } from "./product-price.model";
+
+export type ProductVariantModel = {
+  id?: string;
+  title: string;
+  sku: string;
+  barcode?: string;
+  prices?: ProductPriceModel[];
+  inventoryQuantity?: number;
+  allowBackorder?: boolean;
+};
